Data subsystem use fsspec (#1531)
Delete the fsspec plugin and fix tests.  bring in usage into standard structureddataset encoders/decoders.
