+++ flytekit/core/data_persistence.py
@@ -25,11 +25,10 @@ import os
 import pathlib
 import tempfile
 import typing
-from typing import Union
+from typing import Union, cast
 from uuid import UUID
 
 import fsspec
-from fsspec.core import strip_protocol
 from fsspec.utils import get_protocol
 
 from flytekit import configuration
@@ -95,7 +94,7 @@ class FileAccessProvider(object):
 
         self._data_config = data_config if data_config else DataConfig.auto()
         self._default_protocol = get_protocol(raw_output_prefix)
-        self._default_remote = self.get_filesystem(self._default_protocol)
+        self._default_remote = cast(fsspec.AbstractFileSystem, self.get_filesystem(self._default_protocol))
         if os.name == "nt" and raw_output_prefix.startswith("file://"):
             raise FlyteAssertion("Cannot use the file:// prefix on Windows.")
         self._raw_output_prefix = (
@@ -113,11 +112,11 @@ class FileAccessProvider(object):
         return self._data_config
 
     def get_filesystem(
-        self, protocol: str = None, anonymous: bool = False
+        self, protocol: typing.Optional[str] = None, anonymous: bool = False
     ) -> typing.Optional[fsspec.AbstractFileSystem]:
         if not protocol:
             return self._default_remote
-        kwargs = {}
+        kwargs = {}  # type: typing.Dict[str, typing.Any]
         if protocol == "file":
             kwargs = {"auto_mkdir": True}
         elif protocol == "s3":
@@ -134,9 +133,9 @@ class FileAccessProvider(object):
 
         return fsspec.filesystem(protocol, **kwargs)  # type: ignore
 
-    def get_filesystem_for_path(self, path: str = "") -> fsspec.AbstractFileSystem:
+    def get_filesystem_for_path(self, path: str = "", anonymous: bool = False) -> fsspec.AbstractFileSystem:
         protocol = get_protocol(path)
-        return self.get_filesystem(protocol)
+        return self.get_filesystem(protocol, anonymous=anonymous)
 
     @staticmethod
     def is_remote(path: Union[str, os.PathLike]) -> bool:
@@ -322,7 +321,7 @@ class FileAccessProvider(object):
         """
         try:
             with PerformanceTimer(f"Writing ({local_path} -> {remote_path})"):
-                self.put(local_path, remote_path, recursive=is_multipart)
+                self.put(cast(str, local_path), remote_path, recursive=is_multipart)
         except Exception as ex:
             raise FlyteAssertion(
                 f"Failed to put data from {local_path} to {remote_path} (recursive={is_multipart}).\n\n"
