+++ flytekit/remote/remote.py
@@ -6,17 +6,20 @@ but in Python object form.
 from __future__ import annotations
 
 import base64
-import functools
 import hashlib
+import importlib
 import os
 import pathlib
+import tempfile
 import time
 import typing
 import uuid
+from base64 import b64encode
 from collections import OrderedDict
 from dataclasses import asdict, dataclass
 from datetime import datetime, timedelta
 
+import requests
 from flyteidl.admin.signal_pb2 import Signal, SignalListRequest, SignalSetRequest
 from flyteidl.core import literals_pb2 as literals_pb2
 
@@ -31,10 +34,15 @@ from flytekit.core.data_persistence import FileAccessProvider
 from flytekit.core.launch_plan import LaunchPlan
 from flytekit.core.python_auto_container import PythonAutoContainerTask
 from flytekit.core.reference_entity import ReferenceSpec
+from flytekit.core.tracker import get_full_module_path
 from flytekit.core.type_engine import LiteralsResolver, TypeEngine
 from flytekit.core.workflow import WorkflowBase
 from flytekit.exceptions import user as user_exceptions
-from flytekit.exceptions.user import FlyteEntityAlreadyExistsException, FlyteEntityNotExistException
+from flytekit.exceptions.user import (
+    FlyteEntityAlreadyExistsException,
+    FlyteEntityNotExistException,
+    FlyteValueException,
+)
 from flytekit.loggers import remote_logger
 from flytekit.models import common as common_models
 from flytekit.models import filters as filter_models
@@ -62,7 +70,7 @@ from flytekit.remote.interface import TypedInterface
 from flytekit.remote.lazy_entity import LazyEntity
 from flytekit.remote.remote_callable import RemoteEntity
 from flytekit.tools.fast_registration import fast_package
-from flytekit.tools.script_mode import fast_register_single_script, hash_file
+from flytekit.tools.script_mode import compress_single_script, hash_file
 from flytekit.tools.translator import (
     FlyteControlPlaneEntity,
     FlyteLocalEntity,
@@ -728,7 +736,23 @@ class FlyteRemote(object):
             content_md5=md5_bytes,
             filename=to_upload.name,
         )
-        self._ctx.file_access.put_data(str(to_upload), upload_location.signed_url)
+
+        encoded_md5 = b64encode(md5_bytes)
+        with open(str(to_upload), "+rb") as local_file:
+            content = local_file.read()
+            content_length = len(content)
+            rsp = requests.put(
+                upload_location.signed_url,
+                data=content,
+                headers={"Content-Length": str(content_length), "Content-MD5": encoded_md5},
+            )
+
+            if rsp.status_code != requests.codes["OK"]:
+                raise FlyteValueException(
+                    rsp.status_code,
+                    f"Request to send data {upload_location.signed_url} failed.",
+                )
+
         remote_logger.debug(
             f"Uploading {to_upload} to {upload_location.signed_url} native url {upload_location.native_url}"
         )
@@ -795,16 +819,14 @@ class FlyteRemote(object):
         if image_config is None:
             image_config = ImageConfig.auto_default_image()
 
-        upload_location, md5_bytes = fast_register_single_script(
-            source_path,
-            module_name,
-            functools.partial(
-                self.client.get_upload_signed_url,
-                project=project or self.default_project,
-                domain=domain or self.default_domain,
-                filename="scriptmode.tar.gz",
-            ),
-        )
+        with tempfile.TemporaryDirectory() as tmp_dir:
+            archive_fname = pathlib.Path(os.path.join(tmp_dir, "script_mode.tar.gz"))
+            mod = importlib.import_module(module_name)
+            compress_single_script(source_path, str(archive_fname), get_full_module_path(mod, mod.__name__))
+            md5_bytes, upload_native_url = self._upload_file(
+                archive_fname, project or self.default_project, domain or self.default_domain
+            )
+
         serialization_settings = SerializationSettings(
             project=project,
             domain=domain,
@@ -813,7 +835,7 @@ class FlyteRemote(object):
             fast_serialization_settings=FastSerializationSettings(
                 enabled=True,
                 destination_dir=destination_dir,
-                distribution_location=upload_location.native_url,
+                distribution_location=upload_native_url,
             ),
         )
