+++ plugins/flytekit-data-fsspec/setup.py
@@ -4,7 +4,7 @@ PLUGIN_NAME = "fsspec"
 
 microlib_name = f"flytekitplugins-data-{PLUGIN_NAME}"
 
-plugin_requires = ["flytekit>=1.3.0b2,<2.0.0", "fsspec>=2021.7.0", "botocore>=1.7.48", "pandas>=1.2.0"]
+plugin_requires = []
 
 __version__ = "0.0.0+develop"
 
@@ -13,7 +13,7 @@ setup(
     version=__version__,
     author="flyteorg",
     author_email="admin@flyte.org",
-    description="This package data-plugins for flytekit, that are powered by fsspec",
+    description="This is a deprecated plugin as of flytekit 1.5",
     url="https://github.com/flyteorg/flytekit/tree/master/plugins/flytekit-data-fsspec",
     long_description=open("README.md").read(),
     long_description_content_type="text/markdown",
@@ -22,9 +22,9 @@ setup(
     install_requires=plugin_requires,
     extras_require={
         # https://github.com/fsspec/filesystem_spec/blob/master/setup.py#L36
-        "abfs": ["adlfs>=2022.2.0"],
-        "aws": ["s3fs>=2021.7.0"],
-        "gcp": ["gcsfs>=2021.7.0"],
+        "abfs": [],
+        "aws": [],
+        "gcp": [],
     },
     license="apache2",
     python_requires=">=3.8",
